+++ src/main/java/metty1337/currencyexchange/dto/ExchangeDTO.java
@@ -4,13 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Setter
 @Getter
 public class ExchangeDTO {
     private CurrencyDTO baseCurrency;
     private CurrencyDTO targetCurrency;
-    private double rate;
-    private double amount;
-    private double convertedAmount;
+    private BigDecimal rate;
+    private BigDecimal amount;
+    private BigDecimal convertedAmount;
 }
