+++ src/main/java/metty1337/currencyexchange/models/ExchangeRate.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Setter
 @Getter
@@ -11,5 +13,5 @@ public class ExchangeRate {
     private Integer ID;
     private Integer BaseCurrencyID;
     private Integer TargetCurrencyID;
-    private double Rate;
+    private BigDecimal Rate;
 }
